Move from InstanceExport to CurseForge sub folder. This is for when more export options are added later.
